MODIFY: set try & exception on quiz main api
