+++ questions/serializers.py
@@ -3,7 +3,6 @@ import random
 from rest_framework import serializers
 
 from questions.models import Answer, Category, Question
-from users.models     import UserRank
 
 
 class CategoryListSerializer(serializers.ModelSerializer):
@@ -14,12 +13,6 @@ class CategoryListSerializer(serializers.ModelSerializer):
 
 
 class CategoryDetailSerializer(serializers.ModelSerializer):
-    attempt = serializers.SerializerMethodField()
-
-    def get_attempt(self, object):
-        user    = self.context['request'].user
-        attempt = UserRank.objects.filter(user=user.id)
-        return attempt
 
     class Meta:
         model  = Category
@@ -30,18 +23,21 @@ class QuizSerializer(serializers.ModelSerializer):
     content = serializers.SerializerMethodField()
 
     def get_content(self, object):
-        question_list = list(Question.objects.filter(category=object.id))
-        
-        data  = [{
-                    'question': question.content,
-                    'image'   : question.image,
-                    'answer'  : [{
-                                    "answer"        : answer.correct,
-                                    "answer_content": answer.content
-                                } for answer in random.sample(list(Answer.objects.filter(question=question)), 5)]
-                } for question in random.sample(question_list, 10)]
+        try:
+            question_list = list(Question.objects.filter(category=object.id))
+            
+            data  = [{
+                        'question': question.content,
+                        'image'   : question.image,
+                        'answer'  : [{
+                                        "answer"        : answer.correct,
+                                        "answer_content": answer.content
+                                    } for answer in random.sample(list(Answer.objects.filter(question=question)), 5)]
+                    } for question in random.sample(question_list, 10)]
 
-        return data
+            return data
+        except ValueError:
+            raise serializers.ValidationError("try again")
 
     class Meta:
         model = Category
